Changed the initialization of the test classes so they can run in parallel.
git-svn-id: http://ikube.googlecode.com/svn/trunk@371 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/test/java/ikube/BaseTest.java
@@ -1,5 +1,6 @@
 package ikube;
 
+import ikube.listener.ListenerManager;
 import ikube.model.IndexContext;
 import ikube.model.faq.Attachment;
 import ikube.model.faq.Faq;
@@ -15,6 +16,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.Persistence;
 
 import org.junit.AfterClass;
+import org.junit.BeforeClass;
 
 /**
  * @author Michael Couck
@@ -23,34 +25,40 @@ import org.junit.AfterClass;
  */
 public abstract class BaseTest extends ATest {
 
-	static {
-		try {
-			ApplicationContextManager.getApplicationContext(IConstants.SPRING_CONFIGURATION_FILE);
-			// Delete all the old index directories
-			Map<String, IndexContext> contexts = ApplicationContextManager.getBeans(IndexContext.class);
-			for (IndexContext indexContext : contexts.values()) {
-				File baseIndexDirectory = FileUtilities.getFile(indexContext.getIndexDirectoryPath(), Boolean.TRUE);
-				FileUtilities.deleteFile(baseIndexDirectory, 1);
-			}
-			PerformanceTester.execute(new PerformanceTester.APerform() {
-				@Override
-				public void execute() throws Exception {
-					EntityManager entityManager = null;
-					try {
-						entityManager = Persistence.createEntityManagerFactory(IConstants.PERSISTENCE_UNIT_NAME).createEntityManager();
-						Class<?>[] classes = new Class[] { Faq.class, Attachment.class };
-						DataGeneratorFour dataGenerator = new DataGeneratorFour(entityManager, 10, classes);
-						dataGenerator.before();
-						dataGenerator.generate();
-						dataGenerator.after();
-					} finally {
-						entityManager.close();
-					}
-				}
-			}, "Data generator two insertion : ", 1);
-		} catch (Exception e) {
-			e.printStackTrace();
+	/** We only want the before class to execute once for the whole test batch. */
+	private static boolean INIT = Boolean.FALSE;
+
+	@BeforeClass
+	public static void beforeClass() throws Exception {
+		if (INIT) {
+			return;
 		}
+		INIT = Boolean.TRUE;
+		ApplicationContextManager.getApplicationContext(IConstants.SPRING_CONFIGURATION_FILE);
+		// Delete all the old index directories
+		Map<String, IndexContext> contexts = ApplicationContextManager.getBeans(IndexContext.class);
+		for (IndexContext indexContext : contexts.values()) {
+			File baseIndexDirectory = FileUtilities.getFile(indexContext.getIndexDirectoryPath(), Boolean.TRUE);
+			FileUtilities.deleteFile(baseIndexDirectory, 1);
+		}
+		PerformanceTester.execute(new PerformanceTester.APerform() {
+			@Override
+			public void execute() throws Exception {
+				EntityManager entityManager = null;
+				try {
+					entityManager = Persistence.createEntityManagerFactory(IConstants.PERSISTENCE_UNIT_NAME).createEntityManager();
+					Class<?>[] classes = new Class[] { Faq.class, Attachment.class };
+					DataGeneratorFour dataGenerator = new DataGeneratorFour(entityManager, 10, classes);
+					dataGenerator.before();
+					dataGenerator.generate();
+					dataGenerator.after();
+				} finally {
+					entityManager.close();
+				}
+			}
+		}, "Data generator two insertion : ", 1);
+		// Remove all the listeners as they create havoc
+		ListenerManager.removeListeners();
 	}
 
 	@AfterClass
@@ -61,6 +69,10 @@ public abstract class BaseTest extends ATest {
 			File indexDirectory = FileUtilities.getFile(indexContext.getIndexDirectoryPath(), Boolean.TRUE);
 			FileUtilities.deleteFile(indexDirectory, 1);
 		}
+	}
+
+	public BaseTest(Class<?> subClass) {
+		super(subClass);
 	}
 
 	protected IndexContext indexContext = ApplicationContextManager.getBean("indexContext");
